+++ packages/cli/src/util/oauth.ts
@@ -12,6 +12,7 @@ interface AuthorizationServerMetadata {
   token_endpoint: URL;
   revocation_endpoint: URL;
   jwks_uri: URL;
+  introspection_endpoint: URL;
 }
 
 let _as: AuthorizationServerMetadata;
@@ -63,7 +64,8 @@ async function processDiscoveryEndpointResponse(
     !canParseURL(json.device_authorization_endpoint) ||
     !canParseURL(json.token_endpoint) ||
     !canParseURL(json.revocation_endpoint) ||
-    !canParseURL(json.jwks_uri)
+    !canParseURL(json.jwks_uri) ||
+    !canParseURL(json.introspection_endpoint)
   ) {
     return [new TypeError('Invalid discovery response')];
   }
@@ -84,6 +86,7 @@ async function processDiscoveryEndpointResponse(
       token_endpoint: new URL(json.token_endpoint),
       revocation_endpoint: new URL(json.revocation_endpoint),
       jwks_uri: new URL(json.jwks_uri),
+      introspection_endpoint: new URL(json.introspection_endpoint),
     },
   ];
 }
@@ -397,4 +400,52 @@ function canParseURL(url: string) {
   } catch {
     return false;
   }
+}
+
+/**
+ * Perform Token Introspection Request.
+ *
+ * @see https://datatracker.ietf.org/doc/html/rfc7662#section-2.1
+ */
+export async function inspectTokenRequest(token: string): Promise<Response> {
+  return fetch((await as()).introspection_endpoint, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/x-www-form-urlencoded',
+      'user-agent': ua,
+    },
+    body: new URLSearchParams({ token }),
+  });
+}
+
+/**
+ * @see https://datatracker.ietf.org/doc/html/rfc7662#section-2.2 */
+interface AccessToken {
+  /** Whether or not the presented token is active. */
+  active: boolean;
+  client_id?: string;
+  session_id?: string;
+}
+
+/**
+ * Process Token Introspection Response.
+ *
+ * @see https://datatracker.ietf.org/doc/html/rfc7662#section-2.2
+ */
+export async function processInspectTokenResponse(
+  response: Response
+): Promise<[IntrospectionError] | [null, AccessToken]> {
+  try {
+    const token = await response.json();
+    if (!token || typeof token !== 'object' || !('active' in token)) {
+      throw new IntrospectionError('Invalid token introspection response');
+    }
+    return [null, token];
+  } catch (cause) {
+    return [new IntrospectionError('Could not introspect token.', { cause })];
+  }
+}
+
+class IntrospectionError extends Error {
+  name = 'IntrospectionError';
 }