+++ packages/cli/test/unit/commands/login/future.test.ts
@@ -43,6 +43,7 @@ describe('login --future', () => {
         token_endpoint: 'https://vercel.com',
         revocation_endpoint: 'https://vercel.com',
         jwks_uri: 'https://vercel.com',
+        introspection_endpoint: 'https://vercel.com',
       })
     );
     const _as = await oauth.as();